Refactor: Separate type definitions
